refactor(cli): migrate chc tooling onto discovery
Move CHC's anonymous-FTP refresher/prober/validator plus its ftp_bases curated-id
and bundled-id resolver to earthlens.chc.cli, and generalise core's _BUNDLED_IDS
(the audit diff axis) to role dispatch so core names no backend there either.
Relocate the CLI tests, keeping the mis-grouped NWP availability tests in core.
Shrink the guard allow-list (chc).
